+++ nrv/__init__.py
@@ -33,7 +33,7 @@ U. of Bordeaux.
   https://doi.org/10.1371/journal.pcbi.1011826
 
 
-.. SeeAlso::
+.. seealso::
    - **General information**: `nrv-framework.org <https://nrv-framework.org>`_
    - **Discussions and queries**: `Forum NRV <https://nrv-framework.org/forum>`_
    - **Full code**: `Github repository <https://github.com/nrv-framework/NRV>`_