+++ tests/deprecated_tests/083_test_parallel_conservative_sum.py
@@ -1,3 +1,4 @@
+# deprecated since at least v1.2.2
 import nrv
 import numpy as np