+++ tests/deprecated_tests/163_test_FENICS_FEMsim_myelinated.py
@@ -1,3 +1,4 @@
+# deprecated since at least v1.2.2
 import nrv
 import time
 import os